+++ src/modules/List.js
@@ -1,0 +1,9 @@
+class List {
+  constructor(description, completed, id) {
+    this.description = description;
+    this.completed = completed;
+    this.id = id;
+  }
+}
+
+export default List;
